% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addfiglabel.R
\name{addfiglabel}
\alias{addfiglabel}
\title{add an identifier to subpanels in a plot}
\usage{
addfiglabel(lab, xloc = -1, yloc = 1, ...)
}
\arguments{
\item{lab}{character, the text to appear (e.g. "(a)")}

\item{xloc}{numeric, the location of the text along the x-axis (default: -1)}

\item{yloc}{numeric, the location of the text along the y-axis (default: 1)}

\item{...}{additional parameters passed on the \code{text()} function, e.g. \code{cex}, \code{col} etc}
}
\value{
a plot
}
\description{
add an identifier to subpanels in a plot
}
\details{
the function produces a new plot on top of the current one, albeit without axes and annotations (think of a new 'layer'). The text is then written according to the coordinates passed on, which by default places the text label in the top left corner.
}
\examples{
par(mfcol = c(2, 2))
plot(1:10)
addfiglabel("(a)")
plot(1:10)
addfiglabel("(b)", cex = 0.5, col = "red")
par(mar = c(2,2,2,2))
plot(1:10)
addfiglabel("(c)")
plot(1:10)
addfiglabel("(d)", cex = 3)
}
