% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archie.R
\docType{data}
\name{archie}
\alias{archie}
\title{Example data from Archie et al 2006}
\format{integer matrix}
\usage{
archie
}
\description{
Example data from Archie et al 2006
}
\details{
family CB (figure 2, p. 123)
}
\references{
Archie, E. A., Morrison, T. A., Foley, C. A., Moss, C. J., & Alberts, S. C. (2006). Dominance rank relationships among wild female African elephants, Loxodonta africana. Animal Behaviour, 71(1), 117-127. \url{https://doi.org/10.1016/j.anbehav.2005.03.023}
}
\keyword{datasets}
