% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcranks.R
\name{calcranks}
\alias{calcranks}
\title{add dominance indices to a data set given a dominance matrix}
\usage{
calcranks(mat, respdata, radagio = TRUE)
}
\arguments{
\item{mat}{a dominance matrix, most likely from \code{\link{datagenfunc}}}

\item{respdata}{a \code{data.frame} that contains a \code{ID} column, which in turn contains IDs that appear in \code{mat}}

\item{radagio}{logical, by default \code{TRUE}, see details}
}
\value{
a \code{data.frame} with metrics added
}
\description{
add dominance indices to a data set given a dominance matrix
}
\details{
currently, four metrics are calculated:
\itemize{
  \item David's score (via the EloRating package)
  \item randomized Elo-rating (via the EloChoice package)
  \item I&SI (this package)
  \item ADAGIO (this package)
}

All metrics are returned twice, on their original scale and z-transformed. In addition, the 'true rank' (referred to in the \code{rank} column) is also z-transformed.

This function adds new columns to \code{xdata}. For this to work, \code{xdata} has to have at least two columns with the name \code{ID} and \code{rank}.

If \code{radagio = FALSE}, the function returns ranks as individuals ranked by their proportion of wins. This is meant as a testing strategy because running the ADAGIO algorithm in turn requires the Java ADAGIO software to be run, which first of all is very slow and also interupts workflow because this produces a pop-up window to appear for each instance. The intended use of this argument is to try out simulations that require the \code{calcrank()} function. If the simulation works, then this argument can be set to \code{radagio = TRUE} (its default).
}
\examples{
\dontrun{
domdata <- datagenfunc(gs = 5, Nintm = 3)
xdata <- createresponse(domdata[[2]], 2, effectsize = 2, error = 0.5)
calcranks(domdata[[1]], xdata)
}
}
