% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createresponse.R
\name{createresponse}
\alias{createresponse}
\title{generate a response variable, given ranks and effect size}
\usage{
createresponse(ranks, Nmulti, effectsize, error)
}
\arguments{
\item{ranks}{a named vector with the ranks (and individual IDs as name); typically the output from \code{\link{datagenfunc}}}

\item{Nmulti}{integer, the number of observations to create per individual}

\item{effectsize}{numeric, the effect size}

\item{error}{numeric, the error}
}
\value{
a \code{data.frame} with \code{length(ranks)} * \code{Nmulti} rows
}
\description{
generate a response variable, given ranks and effect size
}
\details{
the function calculates a response variable with the following formula:

resp = \code{effectsize} * \code{ranks} + \code{rnorm(..., mean = 0, sd = error)}

in addition, a random intercept is generated:

random intercept = \code{rnorm(length(ranks))}
}
\examples{
ranks <- 1:4
names(ranks) <- LETTERS[1:4]
xdata <- createresponse(ranks, 2, effectsize = 2, error = 0.5)
plot(xdata$rank, xdata$response)
}
