% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenfunc.R
\name{datagenfunc}
\alias{datagenfunc}
\title{Data generation}
\usage{
datagenfunc(gs, Nintm, mode = "lin", shrinkranks = TRUE,
  allowties = TRUE)
}
\arguments{
\item{gs}{integer, group size}

\item{Nintm}{integer, multiplier for interactions}

\item{mode}{character, either \code{"lin"} for linear mode or \code{"2"}, \code{"3"} or \code{"4"} for exponential modes}

\item{shrinkranks}{logical, should generated ranks be 'shrunk' so that no gaps occur (default is \code{TRUE}), see details}

\item{allowties}{logical, should tied/shared ranks be allowed in the data generation (default is \code{TRUE})}
}
\value{
a list, the first is the dominance matrix, the second are the true ranks (upon which simulation of outcomes of interactions is based)
}
\description{
Generate dominance matrices following the description of Douglas et al 2017.
}
\details{
if \code{shrinkranks = FALSE} rank orders are returned that may have gaps, e.g. an allowed generated order might be 1, 1, 3, 3, 5, whereas 1, 1, 2, 2, 3 would not be allowed in this scenario (but it would be if \code{shrinkranks = TRUE})
}
\examples{
set.seed(123)
datagenfunc(10, 2, shrinkranks = TRUE)[[2]]
# two IDs with rank=3 and one with rank=4
set.seed(123)
datagenfunc(10, 2, shrinkranks = FALSE)[[2]]
# two IDs with rank=3 and none with rank=4, but the next rank would be rank=5
}
