% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnh.R
\docType{data}
\name{dnh}
\alias{dnh}
\title{Example data from Douglas et al 2017}
\format{integer matrix}
\usage{
dnh
}
\description{
Example data from Douglas et al 2017
}
\details{
figure 1, p. 23
}
\references{
Douglas, P. H., Ngomo, A. C. N., & Hohmann, G. (2017). A novel approach for dominance assessment in gregarious species: ADAGIO. Animal Behaviour, 123, 21-32. \url{https://doi.org/10.1016/j.anbehav.2016.10.014}
}
\keyword{datasets}
