% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISI-diffmat.R
\name{.diffmat}
\alias{.diffmat}
\title{difference matrix}
\usage{
.diffmat(mat)
}
\arguments{
\item{mat}{square interaction matrix with winner in rows and losers in columns}
}
\value{
data frame with inconsistencies and their strength
}
\description{
difference matrix
}
\details{
helper function for \code{\link{ISI}}
}
\examples{
data(bonobos, package = "EloRating")
radagio:::.diffmat(bonobos)

}
\author{
Christof Neumann
}
