% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eucdist.R
\name{eucdist}
\alias{eucdist}
\title{mean Euclidean distance of two vectors}
\usage{
eucdist(x, y)
}
\arguments{
\item{x}{vector 1}

\item{y}{vector 2}
}
\value{
numeric
}
\description{
mean Euclidean distance of two vectors
}
\examples{
X <- c(1, 2, 2)
Y <- c(1, 2, 3)
eucdist(X, Y)
}
