% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindex.R
\name{hindex}
\alias{hindex}
\title{linearity index}
\usage{
hindex(interactionmatrix, loops = 1000)
}
\arguments{
\item{interactionmatrix}{the dominance matrix}

\item{loops}{numeric, the number of loops}
}
\value{
a data.frame with the results
}
\description{
linearity index
}
\examples{
data(archie)
hindex(archie)
}
\references{
de Vries, H. (1995). An improved test of linearity in dominance hierarchies containing unknown or tied relationships. Animal Behaviour, 50(5), 1375-1389. \url{https://doi.org/10.1016/0003-3472(95)80053-0}

Landau, H. (1951). On dominance relations and the structure of animal societies: I. Effect of inherent characteristics. The bulletin of mathematical biophysics, 13(1), 1-19. \url{https://doi.org/10.1007/BF02478336}
}
