% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2seq.R
\name{mat2seq}
\alias{mat2seq}
\title{transform matrix into sequence}
\usage{
mat2seq(mat)
}
\arguments{
\item{mat}{a square matrix}
}
\value{
a data.frame
}
\description{
transform matrix into sequence
}
\examples{
data(archie)
mat2seq(archie)
}
