% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunc.R
\name{plotfunc}
\alias{plotfunc}
\title{Plotting simulation results}
\usage{
plotfunc(reslist, ov, metric = c("euc", "cor"), mode = c("ADAGIO",
  "eloRA", "eloRM", "eloO"), ylimits = NULL, linesstyle = FALSE)
}
\arguments{
\item{reslist}{the output of the simulation steps (see vignettes and appendices of manuscript)}

\item{ov}{the output of the simulation steps (see vignettes and appendices of manuscript)}

\item{metric}{character, \code{"euc"} for Euclidean distance or \code{"cor"} for correlation}

\item{mode}{character, which calculations to take for Elo and DS}

\item{ylimits}{numeric of length 2, overrides internal setting for y-axis limits}

\item{linesstyle}{logic, should different methods differ also with regards to line style, by default \code{FALSE}}
}
\value{
a plot
}
\description{
Plotting simulation results
}
\details{
This is a idiosyncratic function that is only usable in the context of the simulations we run here. See the vignettes of the package.

The \code{"mode"} argument determines which software to use for Elo-ratings and DS. The ADAGIO rankings themselves are always taken from the ADAGIO software. The ISI ranks are always taken from a function included in this package.
\itemize{
  \item \code{"ADAGIO"}: Take ranks for Elo and DS from the ADAGIO software.
  \item \code{"eloRA"}: Calculate ranks for Elo and DS from original scores calculated with the \code{EloRating} (for DS) and \code{EloChoice} (for Elo-rating) packages. The ranking is done via the \code{"average"} option in the \code{rank()} function.
  \item \code{"eloRM"}: Same as above, but the ranks are derived in the following way: in case of ties the minimum rank is taken and subsequently ranks are shrunk so that there will be no gaps. See examples in \code{\link{ranks}}.
  \item \code{"eloO"}: Here the original scores for Elo and DS are used, and if \code{metric="cor"} the Pearson correlation is calculated.
}
}
