% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranks.R
\name{ranks}
\alias{ranks}
\title{ranking}
\usage{
ranks(x, tiestreat = c("avg", "minshrink"))
}
\arguments{
\item{x}{numeric, the dominance index}

\item{tiestreat}{either \code{"average"} or \code{"minshrink"}}
}
\value{
a vector with ranks (the same length as \code{x})
}
\description{
return ranks that treat tied values either as average or that use the minimum rank and shrink ranks so that the gap between adjacent ranks is always 1
}
\examples{
x <- c(1, 2, 2, 7, 8)
# maximum rank possible is 5
ranks(x, tiestreat = "avg")
ranks(x, tiestreat = "minshrink")
x <- c(7, 2, 2, 3, 10, 10)
# maximum rank possible is 6
ranks(x, tiestreat = "avg")
ranks(x, tiestreat = "minshrink")
}
