% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shi2.R
\docType{data}
\name{shi2}
\alias{shi2}
\title{Example data from Shimoji et al 2014}
\format{integer matrix}
\usage{
shi2
}
\description{
Example data from Shimoji et al 2014
}
\details{
group C2
}
\references{
Shimoji, H., Abe, M. S., Tsuji, K., & Masuda, N. (2014). Global network structure of dominance hierarchy of ant workers. Journal of the Royal Society Interface, 11(99), 20140599. \url{https://doi.org/10.1098/rsif.2014.0599}
}
\keyword{datasets}
